+++ lib/computed/computed.dart
@@ -1,0 +1,8 @@
+
+class Computed<S, T> {
+  final T Function(S state, T previousValue) _closure;
+
+  Computed(this._closure);
+
+  T call(S state, T previousValue) => _closure(state, previousValue);
+}