+++ lib/core/middleware.dart
@@ -1,7 +1,0 @@
-import 'package:meowchannel/core/dispatcher.dart';
-
-typedef Middleware<S> = Dispatcher Function(
-  Dispatcher dispatcher,
-  S Function() getState,
-  Dispatcher next
-);