+++ lib/extensions/flutter/logger/store_logger_module.dart
@@ -1,0 +1,10 @@
+
+import 'package:flutter/foundation.dart';
+import 'package:meowchannel/core/module.dart';
+
+Module<S> storeLoggerModule<S>(String tag) => Module<S>('storeLoggerModule', (dispatcher, initialize, state, next) {
+  return (dynamic action) async {
+    if(kDebugMode) print('[$tag] $state <- $action');
+    return next(action);
+  };
+});