+++ lib/extensions/flutter/store_hook.dart
@@ -3,7 +3,7 @@ import 'package:meowchannel/utils/type_utils.dart';
 
 class StoreHook<S> {
 
-  final Function(Store<S>, S) _closure;
+  final Function(Store<S> store, S state, dynamic action) _closure;
 
   StoreHook(this._closure):
     assert(_closure != null);
@@ -11,5 +11,5 @@ class StoreHook<S> {
   String get storeType => typeOf<Store<S>>().toString();
   String get stateType => typeOf<S>().toString();
 
-  void apply(Store<S> store, S state) => _closure(store, state);
+  void apply(Store<S> store, S state, dynamic action) => _closure(store, state, action);
 }