+++ lib/worker/worker_context.dart
@@ -10,40 +10,20 @@ class WorkerContext<S> {
     this._state
   );
 
-  Future<T> compute0<T>(Future<T> Function() computable) async {
-    return Executor().execute(priority: WorkPriority.immediately, arg1: <dynamic>[], fun1: (List<dynamic> params) async {
-      return await computable();
-    });
-  }
-
   Future<T> compute1<A, T>(Future<T> Function(A) computable, A a) async {
-    return Executor().execute(priority: WorkPriority.immediately, arg1: <dynamic>[a], fun1: (List<dynamic> params) async {
-      return await computable(params[0] ?? null);
-    });
+    return Executor().execute(priority: WorkPriority.immediately, arg1: a, fun1: computable);
   }
 
   Future<T> compute2<A, B, T>(Future<T> Function(A, B) computable, A a, B b) async {
-    return Executor().execute(priority: WorkPriority.immediately, arg1: <dynamic>[a, b], fun1: (List<dynamic> params) async {
-      return await computable(params[0] ?? null, params[1] ?? null);
-    });
+    return Executor().execute(priority: WorkPriority.immediately, arg1: a, arg2: b, fun2: computable);
   }
 
   Future<T> compute3<A, B, C, T>(Future<T> Function(A, B, C) computable, A a, B b, C c) async {
-    return Executor().execute(priority: WorkPriority.immediately, arg1: <dynamic>[a, b, c], fun1: (List<dynamic> params) async {
-      return await computable(params[0] ?? null, params[1] ?? null, params[2] ?? null);
-    });
+    return Executor().execute(priority: WorkPriority.immediately, arg1: a, arg2: b, arg3: c, fun3: computable);
   }
 
   Future<T> compute4<A, B, C, D, T>(Future<T> Function(A, B, C, D) computable, A a, B b, C c, D d) async {
-    return Executor().execute(priority: WorkPriority.immediately, arg1: <dynamic>[a, b, c, d], fun1: (List<dynamic> params) async {
-      return await computable(params[0] ?? null, params[1] ?? null, params[2] ?? null, params[3] ?? null);
-    });
-  }
-
-  Future<T> compute5<A, B, C, D, E, T>(Future<T> Function(A, B, C, D, E) computable, A a, B b, C c, D d, E e) async {
-    return Executor().execute(arg1: <dynamic>[a, b, c, d, e], fun1: (List<dynamic> params) async {
-      return await computable(params[0] ?? null, params[1] ?? null, params[2] ?? null, params[3] ?? null, params[4] ?? null);
-    });
+    return Executor().execute(priority: WorkPriority.immediately, arg1: a, arg2: b, arg3: c, arg4: d, fun4: computable);
   }
 
   S state() {